[ci skip] Merge metadata and representation in SkoreEstimatorReportItem
