--- skore-remote-project/src/skore_remote_project/item/skore_estimator_report_item/__init__.py
@@ -1,3 +1,0 @@
-from .skore_estimator_report_item import SkoreEstimatorReportItem
-
-__all__ = ["SkoreEstimatorReportItem"]
